requirements.txt
streamlit>=1.31.0
langchain>=0.1.9
langchain-ollama>=0.1.0
langchain-google-genai>=1.0.0
langchain-community>=0.0.20
faiss-cpu>=1.7.4
pandas>=2.2.0
transformers>=4.30.0
torch>=2.0.0
kaggle>=1.5.16
kagglehub>=0.2.0
python-dotenv>=1.0.0

